rename files to lower-case
