+++ src/components/auth.js
@@ -1,5 +1,5 @@
 import {useState} from 'react'
-import {supabase} from '../utils/supabaseClient'
+import {supabase} from '../utils/supabase-client'
 
 export default function Auth() {
 	const [loading, setLoading] = useState(false)
