+++ src/components/nav.js
@@ -1,16 +1,29 @@
 import {Link} from 'react-router-dom'
-import {supabase} from '../utils/supabaseClient'
+import {DbSessionContext} from '../contexts/db-session'
 import ThemeToggleButton from './theme-toggle-button'
 
 export default function Nav() {
 	return (
-		<nav className="Nav">
-			<Link to="/">Home</Link>
-			<Link to="/test">Test</Link>
-			<ThemeToggleButton></ThemeToggleButton>
-			<button className="Nav-signout" onClick={() => supabase.auth.signOut()}>
-				Sign Out
-			</button>
-		</nav>
+		<DbSessionContext.Consumer>
+			{({session}) => {
+				return (
+					<nav className="Nav">
+						<Link to="/">Home</Link>
+						{session ? (
+							<>
+								<Link to="/account">Account</Link>
+								<Link to="/logout">Logout</Link>
+							</>
+						) : (
+							<>
+								<Link to="/login">Login</Link>
+								<Link to="/register">Register</Link>
+							</>
+						)}
+						<ThemeToggleButton></ThemeToggleButton>
+					</nav>
+				)
+			}}
+		</DbSessionContext.Consumer>
 	)
 }
