+++ src/contexts/db-session.js
@@ -1,0 +1,13 @@
+import {createContext} from 'react'
+
+const dbSessionContextDefault = {
+	session: null,
+	logout: () => {},
+}
+
+const DbSessionContext = createContext(dbSessionContextDefault)
+
+export {
+	DbSessionContext,
+	dbSessionContextDefault
+}
