--- src/contexts/session.js
@@ -1,13 +1,0 @@
-import {createContext} from 'react'
-
-const sessionContextDefault = {
-	session: null,
-	logout: () => {},
-}
-
-const SessionContext = createContext(sessionContextDefault)
-
-export {
-	SessionContext,
-	sessionContextDefault
-}
