+++ src/pages/login.js
@@ -1,9 +1,9 @@
-import {SessionContext} from '../contexts/session'
+import {DbSessionContext} from '../contexts/db-session'
 import Auth from '../components/auth'
 
 export default function PageLogin() {
 	return (
-		<SessionContext.Consumer>
+		<DbSessionContext.Consumer>
 			{({session, logout}) => {
 				return session ? (
 					<div>
@@ -14,6 +14,6 @@ export default function PageLogin() {
 					<Auth></Auth>
 				)
 			}}
-		</SessionContext.Consumer>
+		</DbSessionContext.Consumer>
 	)
 }
