+++ src/pages/logout.js
@@ -1,17 +1,15 @@
 import {Redirect} from 'react-router-dom'
-import {SessionContext} from '../contexts/session'
+import {DbSessionContext} from '../contexts/db-session'
 
 export default function PageLogout() {
 	return (
-		<SessionContext.Consumer>
-			{(value) => {
-				const {session, logout} = value
-				console.log(value, session, logout)
+		<DbSessionContext.Consumer>
+			{({session, logout}) => {
 				if (session) {
 					logout()
 					return <Redirect to='/login'/>
 				}
 			}}
-		</SessionContext.Consumer>
+		</DbSessionContext.Consumer>
 	)
 }
